Omaha #3477 changed plot info locations to floats

Former-commit-id: 26d7336c938550bead4c76cd3d915d42815f705d [formerly 73b2585d802d02fd9e20e2acbbf79e39ca5476b3] [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426] [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d]] [formerly 26d7336c938550bead4c76cd3d915d42815f705d [formerly 73b2585d802d02fd9e20e2acbbf79e39ca5476b3] [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426] [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d]] [formerly 8738e9cecec47e090abd6822026187e958b357f7 [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d] [formerly 8738e9cecec47e090abd6822026187e958b357f7 [formerly 00504d6aa7fefd7b6d28701aae777c0fb7d5ace5]]]]]
Former-commit-id: 8738e9cecec47e090abd6822026187e958b357f7
Former-commit-id: 0b4d05df665d264c74cd08fdebc79ca2bc0256d6 [formerly 393659ece82f58e3e33b9925c3638c440298b48e] [formerly 90a17fa5a6b0c35df35802357453819680aac2f9] [formerly 89dea83e419cc7bed9293974aa02eb361cb96ec6 [formerly cc959db02bdac9bc65cd7a6ddad09207c1ce1224] [formerly 90a17fa5a6b0c35df35802357453819680aac2f9 [formerly a180c212c921a6a0bbeb490045c05185d2207ef3]]]
Former-commit-id: a6fdd44e34b63a2a8cdacd9b959b0301e6411888 [formerly c98315075490f91f8cf521ed63c62b0940c10bfc] [formerly 8b8ad6e45e070c04f3394c37f2e8096340f841ac [formerly fd5c0ea48158e90826270be92bad5eae3b9ca93e]]
Former-commit-id: 8b8ad6e45e070c04f3394c37f2e8096340f841ac
Former-commit-id: 78f80b3057331047de82bff1766d28cf0529acdd

+++ cave/com.raytheon.viz.pointdata/src/com/raytheon/viz/pointdata/rsc/retrieve/FullDataPlotInfoRetriever.java
@@ -15,8 +15,8 @@ import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 
-import com.raytheon.uf.common.inventory.exception.DataCubeException;
 import com.raytheon.uf.common.dataquery.requests.RequestConstraint;
+import com.raytheon.uf.common.inventory.exception.DataCubeException;
 import com.raytheon.uf.common.pointdata.PointDataContainer;
 import com.raytheon.uf.common.pointdata.PointDataView;
 import com.raytheon.uf.common.status.IUFStatusHandler;
@@ -40,6 +40,7 @@ import com.raytheon.viz.pointdata.PlotInfo;
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * Jan 12, 2011            bsteffen     Initial creation
+ * Aug 08, 2014 3477       bclement     changed plot info locations to floats
  * 
  * </pre>
  * 
@@ -121,8 +122,8 @@ public class FullDataPlotInfoRetriever extends AbstractPlotInfoRetriever {
                     if (stationId != null) {
                         info.stationId = pdv.getString(stationId);
                     }
-                    info.latitude = pdv.getNumber(latitude).doubleValue();
-                    info.longitude = pdv.getNumber(longitude).doubleValue();
+                    info.latitude = pdv.getNumber(latitude).floatValue();
+                    info.longitude = pdv.getNumber(longitude).floatValue();
                     info.dataURI = pdv.getString("dataURI");
                     Date vTime = new Date(pdv.getLong(validTime));
                     if (fcstTime != null) {
